Move CLBlast implementation to separate file
Add buffer reuse code (adapted from slaren's cuda implementation)
